requirements.txt
aiosmtplib==4.0.0
APScheduler==3.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.3
nvidia-nccl-cu12==2.25.1
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.5.0
twilio==7.7.0
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
Werkzeug==2.2.2
xgboost==2.1.4
